refactor: with clippy
Former-commit-id: c57ef7336468ec1b2386a1c3faa753468b45eeb3
Former-commit-id: a98ec18b83a192c371cedbf3cd0efbb1586c4552

+++ arm7tdmi/src/disass.rs
@@ -42,7 +42,7 @@ where
         let mut line = String::new();
 
         let addr = self.base + self.pos as Addr;
-        let decoded: D = D::decode_from_bytes(&self.bytes[(self.pos as usize)..], addr);
+        let decoded: D = D::decode_from_bytes(&self.bytes[(self.pos)..], addr);
         let decoded_raw = decoded.get_raw();
         self.pos += self.word_size;
         write!(&mut line, "{addr:8x}:\t{decoded_raw:08x} \t{decoded}").unwrap();
