+++ core/src/bios.rs
@@ -57,7 +57,7 @@ impl BusIO for Bios {
     #[inline]
     fn read_16(&mut self, addr: Addr) -> u16 {
         if self.read_allowed() {
-            self.rom.read_16(addr) as u16
+            self.rom.read_16(addr)
         } else {
             (self.last_opcode >> ((addr & 2) << 3)) as u16
         }
